refactor(connector-tester): extract HeapSampler from VerificationEngine
